@@ -1,156 +1,132 @@
 part of 'core_model.dart';
 
-enum RecoveryGroupType { devices, fiduciaries }
+class RecoveryGroupModel extends Serializable {
+  static const currentVersion = 2;
+  static const boxName = 'groups';
+  static const typeId = 20;
 
-@immutable
-class RecoveryGroupModel extends Equatable {
-  static const currentVersion = 1;
-
+  final int version;
   final GroupId id;
-  final String name;
-  final RecoveryGroupType type;
   final int maxSize;
-  final bool hasSecret;
-  final bool isRestoring;
-  final Map<PeerId, GuardianModel> guardians;
+  final int threshold;
+  final PeerId ownerId;
+  final Map<PeerId, String> guardians;
+  final Map<SecretId, String> secrets;
+
+  RecoveryGroupModel({
+    this.version = currentVersion,
+    GroupId? id,
+    PeerId? ownerId,
+    this.maxSize = 0,
+    this.threshold = 0,
+    this.guardians = const {},
+    this.secrets = const {},
+  })  : id = id ?? GroupId(),
+        ownerId = ownerId ?? PeerId();
 
   @override
   List<Object> get props => [id];
 
-  int get currentSize => guardians.length;
-
-  int get threshold => maxSize == 5 ? 3 : 2;
-
-  int get neededMore => maxSize - currentSize;
-
-  bool get hasMinimal => guardians.length >= threshold;
-
-  bool get isFull => guardians.length == maxSize;
-
-  bool get isNotFull => guardians.length < maxSize;
-
-  bool get isMissed => guardians.length < threshold;
-
-  bool get isNotRestoring => !isRestoring;
-
-  bool get canAddGuardian => !hasSecret && !isRestoring && !isFull;
-
-  bool get canAddSecret => !hasSecret && isFull && !isRestoring;
-
-  bool get canRecoverSecret => hasSecret && hasMinimal;
-
-  const RecoveryGroupModel({
-    required this.id,
-    required this.name,
-    this.type = RecoveryGroupType.devices,
-    this.maxSize = 3,
-    this.hasSecret = false,
-    this.isRestoring = false,
-    this.guardians = const {},
-  }) : assert(name != '');
-
-  factory RecoveryGroupModel.fromBytes(Uint8List value) {
-    final u = Unpacker(value);
-    if (u.unpackInt() != currentVersion) throw const FormatException();
-    return RecoveryGroupModel(
-      id: GroupId(value: Uint8List.fromList(u.unpackBinary())),
-      name: u.unpackString()!,
-      type: RecoveryGroupType.values[u.unpackInt()!],
-      maxSize: u.unpackInt()!,
-      hasSecret: u.unpackBool()!,
-      isRestoring: u.unpackBool()!,
-      guardians: Map.fromEntries(u
-          .unpackList()
-          .map<GuardianModel>((e) =>
-              GuardianModel.fromBytes(Uint8List.fromList(e as List<int>)))
-          .map((g) => MapEntry(g.peerId, g))),
-    );
-  }
-
-  Uint8List toBytes() {
-    final guardiansAsBytes = guardians.values.map((e) => e.toBytes());
-    final p = Packer()
-      ..packInt(currentVersion)
-      ..packBinary(id.value)
-      ..packString(name)
-      ..packInt(type.index)
-      ..packInt(maxSize)
-      ..packBool(hasSecret)
-      ..packBool(isRestoring)
-      ..packListLength(guardiansAsBytes.length);
-    guardiansAsBytes.forEach(p.packBinary);
-    return p.takeBytes();
-  }
-
-  RecoveryGroupModel addGuardian(GuardianModel guardian) {
-    if (isFull) throw RecoveryGroupGuardianLimitexhausted();
-    if (guardians.containsKey(guardian.peerId)) {
-      throw RecoveryGroupGuardianAlreadyExists();
-    }
-    return RecoveryGroupModel(
-      id: id,
-      name: name,
-      type: type,
-      maxSize: maxSize,
-      guardians: {...guardians, guardian.peerId: guardian},
-      isRestoring: isFull ? false : isRestoring,
-      hasSecret: hasSecret,
-    );
-  }
-
-  RecoveryGroupModel completeGroup() => RecoveryGroupModel(
-        id: id,
-        name: name,
-        type: type,
-        maxSize: maxSize,
-        guardians: guardians,
-        isRestoring: isRestoring,
-        hasSecret: true,
-      );
-}
-
-class RecoveryGroupGuardianAlreadyExists implements Exception {
-  static const description =
-      'Guardian with given name already exists in that group!';
-}
-
-class RecoveryGroupGuardianLimitexhausted implements Exception {
-  static const description = 'Guardian group size limit exhausted!';
-}
-
-@immutable
-class GuardianModel extends Equatable {
-  static const currentVersion = 1;
-
-  final PeerId peerId;
-  final String name;
-  final String tag;
-
-  const GuardianModel({
-    required this.peerId,
-    required this.name,
-    this.tag = '',
-  }) : assert(name != '');
+  @override
+  bool get isEmpty => guardians.isEmpty;
 
   @override
-  List<Object> get props => [peerId];
+  bool get isNotEmpty => guardians.isNotEmpty;
 
-  factory GuardianModel.fromBytes(Uint8List value) {
-    final u = Unpacker(value);
-    if (u.unpackInt() != currentVersion) throw const FormatException();
-    return GuardianModel(
-      peerId: PeerId(value: Uint8List.fromList(u.unpackBinary())),
-      name: u.unpackString()!,
-      tag: u.unpackString()!,
-    );
+  String get aKey => id.asKey;
+
+  int get size => guardians.length;
+
+  int get redudancy => maxSize - threshold;
+
+  bool get isFull => guardians.length == maxSize;
+  bool get isNotFull => !isFull;
+
+  bool get isRestoring => isNotFull && secrets.isNotEmpty;
+  bool get isNotRestoring => !isRestoring;
+
+  factory RecoveryGroupModel.fromBytes(List<int> value) {
+    final u = Unpacker(value is Uint8List ? value : Uint8List.fromList(value));
+    final version = u.unpackInt()!;
+    switch (version) {
+      case 1:
+        final groupId = GroupId(
+          token: Uint8List.fromList(u.unpackBinary()),
+          name: u.unpackString()!,
+        );
+        u.unpackInt(); // RecoveryGroupType
+        final maxSize = u.unpackInt()!;
+        final hasSecret = u.unpackBool()!;
+        u.unpackBool()!; // isRestoring
+        final guardians =
+            u.unpackList().map<PeerId>((e) => PeerId.fromBytes(e as List<int>));
+        return RecoveryGroupModel(
+          version: version,
+          id: groupId,
+          ownerId: PeerId(),
+          maxSize: maxSize,
+          threshold: maxSize == 3 ? 2 : 5,
+          guardians: {for (final guardian in guardians) guardian: ''},
+          secrets: {
+            if (hasSecret)
+              SecretId(
+                token: Uint8List(SecretId.size),
+                name: groupId.name,
+              ): ''
+          },
+        );
+      case 2:
+        return RecoveryGroupModel(
+          version: version,
+          id: GroupId.fromBytes(u.unpackBinary()),
+          maxSize: u.unpackInt()!,
+          threshold: u.unpackInt()!,
+          ownerId: PeerId.fromBytes(u.unpackBinary()),
+          guardians: u.unpackMap().map<PeerId, String>((k, v) =>
+              MapEntry(PeerId.fromBytes(k as List<int>), v as String)),
+          secrets: u.unpackMap().map<SecretId, String>((k, v) =>
+              MapEntry(SecretId.fromBytes(k as List<int>), v as String)),
+        );
+
+      default:
+        throw const FormatException(
+          'Unsupported version of RecoveryGroupModel!',
+        );
+    }
   }
 
+  @override
   Uint8List toBytes() {
     final p = Packer()
       ..packInt(currentVersion)
-      ..packBinary(peerId.value)
-      ..packString(name)
-      ..packString(tag);
+      ..packBinary(id.toBytes())
+      ..packInt(maxSize)
+      ..packInt(threshold)
+      ..packBinary(ownerId.toBytes());
+    p.packMapLength(guardians.length);
+    for (final e in guardians.entries) {
+      p.packBinary(e.key.toBytes());
+      p.packString(e.value);
+    }
+    p.packMapLength(secrets.length);
+    for (final e in secrets.entries) {
+      p.packBinary(e.key.toBytes());
+      p.packString(e.value);
+    }
     return p.takeBytes();
   }
+
+  RecoveryGroupModel copyWith({
+    PeerId? ownerId,
+    Map<PeerId, String>? guardians,
+    Map<SecretId, String>? secrets,
+  }) =>
+      RecoveryGroupModel(
+        id: id,
+        ownerId: ownerId ?? this.ownerId,
+        maxSize: maxSize,
+        threshold: threshold,
+        guardians: guardians ?? this.guardians,
+        secrets: secrets ?? this.secrets,
+      );
 }
