@@ -1,86 +1,107 @@
 part of 'core_model.dart';
 
-@immutable
-class SecretShardModel extends Equatable {
-  static const currentVersion = 1;
+class SecretShardModel extends Serializable {
+  static const currentVersion = 2;
+  static const boxName = 'shards';
+  static const typeId = 10;
 
-  final String value;
-  final String ownerName;
+  final int version;
+  final SecretId id;
   final PeerId ownerId;
-  final String groupName;
   final GroupId groupId;
   final int groupSize;
   final int groupThreshold;
+  final String shard;
+
+  const SecretShardModel({
+    this.version = currentVersion,
+    required this.id,
+    required this.ownerId,
+    required this.groupId,
+    required this.groupSize,
+    required this.groupThreshold,
+    required this.shard,
+  });
 
   @override
-  List<Object> get props => [groupId, ownerId];
+  List<Object> get props => [id];
 
-  String get asKey => base64UrlEncode(groupId.value);
+  @override
+  bool get isEmpty => shard.isEmpty;
 
-  bool get isEmpty =>
-      ownerId.isEmpty &&
-      groupId.isEmpty &&
-      ownerName.isEmpty &&
-      groupName.isEmpty &&
-      value.isEmpty;
+  @override
+  bool get isNotEmpty => shard.isNotEmpty;
 
-  SecretShardModel({
-    this.value = '',
-    PeerId? ownerId,
-    this.ownerName = '',
-    GroupId? groupId,
-    this.groupName = '',
-    this.groupSize = 0,
-    this.groupThreshold = 0,
-  })  : ownerId = ownerId ?? PeerId.empty(),
-        groupId = groupId ?? GroupId.empty();
+  String get aKey => id.asKey;
 
-  factory SecretShardModel.fromBytes(Uint8List bytes) {
-    final u = Unpacker(bytes);
-    if (u.unpackInt() != currentVersion) throw const FormatException();
-    return SecretShardModel(
-      value: u.unpackString()!,
-      ownerId: PeerId(value: Uint8List.fromList(u.unpackBinary())),
-      ownerName: u.unpackString()!,
-      groupId: GroupId(value: Uint8List.fromList(u.unpackBinary())),
-      groupName: u.unpackString()!,
-      groupSize: u.unpackInt()!,
-      groupThreshold: u.unpackInt()!,
-    );
+  factory SecretShardModel.fromBytes(List<int> bytes) {
+    final u = Unpacker(bytes is Uint8List ? bytes : Uint8List.fromList(bytes));
+    final version = u.unpackInt()!;
+    switch (version) {
+      case 1:
+        final shard = u.unpackString()!;
+        final ownerId = PeerId(
+          token: Uint8List.fromList(u.unpackBinary()),
+          name: u.unpackString()!,
+        );
+        final groupId = GroupId(
+          token: Uint8List.fromList(u.unpackBinary()),
+          name: u.unpackString()!,
+        );
+        final groupSize = u.unpackInt()!;
+        final groupThreshold = u.unpackInt()!;
+        return SecretShardModel(
+          version: version,
+          id: SecretId(token: Uint8List(SecretId.size), name: groupId.name),
+          shard: shard,
+          ownerId: ownerId,
+          groupId: groupId,
+          groupSize: groupSize,
+          groupThreshold: groupThreshold,
+        );
+
+      case 2:
+        return SecretShardModel(
+          version: version,
+          id: SecretId.fromBytes(u.unpackBinary()),
+          ownerId: PeerId.fromBytes(u.unpackBinary()),
+          groupId: GroupId.fromBytes(u.unpackBinary()),
+          groupSize: u.unpackInt()!,
+          groupThreshold: u.unpackInt()!,
+          shard: u.unpackString()!,
+        );
+
+      default:
+        throw const FormatException('Unsupported version of SecretShardModel');
+    }
   }
 
+  @override
   Uint8List toBytes() {
     final p = Packer()
       ..packInt(currentVersion)
-      ..packString(value)
-      ..packBinary(ownerId.value)
-      ..packString(ownerName)
-      ..packBinary(groupId.value)
-      ..packString(groupName)
+      ..packBinary(id.toBytes())
+      ..packBinary(ownerId.toBytes())
+      ..packBinary(groupId.toBytes())
       ..packInt(groupSize)
-      ..packInt(groupThreshold);
+      ..packInt(groupThreshold)
+      ..packString(shard);
     return p.takeBytes();
   }
 
   SecretShardModel copyWith({
-    String? value,
     PeerId? ownerId,
-    String? ownerName,
-    GroupId? groupId,
-    String? groupName,
-    int? groupSize,
-    int? groupThreshold,
+    String? shard,
   }) =>
       SecretShardModel(
-        value: value ?? this.value,
+        id: id,
         ownerId: ownerId ?? this.ownerId,
-        ownerName: ownerName ?? this.ownerName,
-        groupId: groupId ?? this.groupId,
-        groupName: groupName ?? this.groupName,
-        groupSize: groupSize ?? this.groupSize,
-        groupThreshold: groupThreshold ?? this.groupThreshold,
+        groupId: groupId,
+        groupSize: groupSize,
+        groupThreshold: groupThreshold,
+        shard: shard ?? this.shard,
       );
 
   @override
-  String toString() => '$groupName of $ownerName}';
+  String toString() => '${groupId.name} of ${ownerId.name}';
 }
