@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:package_info_plus/package_info_plus.dart';
 
-import '/src/core/theme_data.dart';
+import '/src/core/theme/theme.dart';
 import '/src/core/widgets/icon_of.dart';
 import '/src/core/di_container.dart';
 import '/src/core/model/core_model.dart';
@@ -16,30 +16,23 @@ class HiddenSettingsWidget extends StatefulWidget {
 class _HiddenSettingsWidgetState extends State<HiddenSettingsWidget> {
   late DIContainer _diContainer;
   int _counter = 0;
-  bool _isVisible = false;
 
   @override
   void initState() {
-    _diContainer = context.read<DIContainer>();
     super.initState();
+    _diContainer = context.read<DIContainer>();
   }
 
   @override
   Widget build(BuildContext context) => GestureDetector(
         behavior: HitTestBehavior.translucent,
-        onTap: () {
-          _counter++;
-          if (_counter > 5) {
-            _counter = 0;
-            setState(() => _isVisible = true);
-          }
-        },
+        onTap: () => _counter++ == 5 ? setState(() {}) : null,
         child: Visibility(
           maintainSize: true,
           maintainAnimation: true,
           maintainState: true,
           maintainInteractivity: false,
-          visible: _isVisible,
+          visible: _counter > 5,
           child: Card(
             child: Column(
               children: [
@@ -52,7 +45,8 @@ class _HiddenSettingsWidgetState extends State<HiddenSettingsWidget> {
                       secondary: const IconOf.splitAndShare(),
                       title: const Text('Proxy connection'),
                       subtitle: const Text(
-                          'Connect through Keyper-operated proxy server'),
+                        'Connect through Keyper-operated proxy server',
+                      ),
                       value: boxSettings.isProxyEnabled,
                       onChanged: (isEnabled) {
                         _diContainer.boxSettings.isProxyEnabled = isEnabled;
@@ -70,13 +64,16 @@ class _HiddenSettingsWidgetState extends State<HiddenSettingsWidget> {
                 Padding(
                   padding: paddingV12 + paddingH20,
                   child: FutureBuilder(
-                      future: PackageInfo.fromPlatform(),
-                      builder: (
-                        context,
-                        AsyncSnapshot<PackageInfo?> snapshot,
-                      ) =>
-                          Text(
-                              'Version:  ${snapshot.data?.version}+${snapshot.data?.buildNumber}')),
+                    future: PackageInfo.fromPlatform(),
+                    builder: (
+                      context,
+                      AsyncSnapshot<PackageInfo?> snapshot,
+                    ) =>
+                        Text(
+                      'Version:  ${snapshot.data?.version}'
+                      '+${snapshot.data?.buildNumber}',
+                    ),
+                  ),
                 ),
               ],
             ),
