@@ -1,5 +1,7 @@
 import 'package:flutter/material.dart';
 
+import '/src/core/model/core_model.dart';
+
 import '../add_guardian_controller.dart';
 import '../../widgets/get_qrcode_widget.dart';
 
@@ -8,6 +10,10 @@ class ScanQRCodePage extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) => GetQRCodeWidget(
-        resultCallback: context.read<AddGuardianController>().setQRCode,
+        resultCallback: (qrCode) {
+          if (qrCode.code == MessageCode.createGroup) {
+            context.read<AddGuardianController>().qrCode = qrCode;
+          }
+        },
       );
 }
