@@ -1,83 +1,55 @@
-import 'dart:async';
 import 'package:sss256/sss256.dart';
 
 import '/src/core/model/core_model.dart';
+import '/src/core/service/analytics_service.dart';
 
 import '../recovery_group_controller.dart';
 
 export 'package:provider/provider.dart';
 
-class AddSecretController extends RecoveryGroupController {
-  final Map<PeerId, SecretShardModel> _shards = {};
-  final Map<PeerId, MessageStatus> _responses = {};
-  final Map<PeerId, StreamSubscription> _subscriptions = {};
-  final Map<PeerId, bool> _statuses = {};
-  final GroupId groupId;
+class AddSecretController extends RecoveryGroupSecretController {
+  String secretName = '';
   String secret = '';
-  Timer? _timer;
-  StreamSubscription<MessageModel>? _subscription;
 
   AddSecretController({
     required super.diContainer,
     required super.pagesCount,
-    required this.groupId,
-  });
-
-  Map<PeerId, bool> get statuses => _statuses;
-
-  Map<PeerId, MessageStatus> get responses => _responses;
-
-  RecoveryGroupModel get group => getGroupById(groupId)!;
-
-  @override
-  void dispose() {
-    for (var s in _subscriptions.values) {
-      s.cancel();
-    }
-    _subscription?.cancel();
-    _timer?.cancel();
-    super.dispose();
-  }
+    required super.groupId,
+  }) : super(secretId: SecretId(name: ''));
 
   void startRequest({
-    Callback? onSuccess,
-    Callback? onReject,
-    Callback? onFailed,
+    required Callback onSuccess,
+    required Callback onReject,
+    required Callback onFailed,
   }) {
-    _subscription = networkStream.listen((MessageModel message) {
-      if (message.type != OperationType.setShard) return;
-      if (!message.hasResponse) return;
-      _responses[message.peerId] = message.status;
-      notifyListeners();
-      if (message.isAccepted) {
-        removeShard(message.peerId);
-        if (_shards.isEmpty) onSuccess?.call(message);
-        return;
-      }
-      if (message.isRejected) {
-        _timer?.cancel();
-        _subscription?.cancel();
-        _shards.clear();
-        onReject?.call(message);
-        return;
-      }
-      onFailed?.call(message);
-    });
-    for (var peerId in group.guardians.keys) {
-      _subscriptions[peerId] = diContainer.networkService.onPeerStatusChanged(
-        (bool isOnline) {
-          if (isOnline == _statuses[peerId]) return;
-          _statuses[peerId] = isOnline;
-          notifyListeners();
-        },
-        peerId,
-      );
-    }
+    diContainer.analyticsService.logEvent(eventStartAddSecret);
+    networkSubscription = networkStream.listen(
+      (message) {
+        if (message.code != MessageCode.setShard) return;
+        if (!message.hasResponse) return;
+        if (!messages.contains(message)) return;
+        updateMessage(message);
+        if (message.isAccepted) {
+          if (messagesWithSuccess.length == group.maxSize) {
+            diContainer.analyticsService.logEvent(eventFinishAddSecret);
+            diContainer.boxRecoveryGroups
+                .put(
+                  group.aKey,
+                  group.copyWith(secrets: {...group.secrets, secretId: ''}),
+                )
+                .then((_) => onSuccess(message));
+          }
+        } else {
+          stopListenResponse();
+          message.isRejected ? onReject(message) : onFailed(message);
+        }
+      },
+    );
     _splitSecret();
-    _timer = Timer.periodic(globals.retryNetworkTimeout, distributeShards);
-    distributeShards();
+    startNetworkRequest(requestShards);
   }
 
+  /// fill messages with request
   void _splitSecret() {
     final shards = splitSecret(
       treshold: group.threshold,
@@ -85,43 +57,25 @@ class AddSecretController extends RecoveryGroupController {
       secret: secret,
     );
     if (secret != restoreSecret(shares: shards.sublist(0, group.threshold))) {
-      throw Exception('Can not restore the secret');
+      throw const FormatException('Can not restore the secret!');
     }
+    secretId = secretId.copyWith(name: secretName);
     final shardsIterator = shards.iterator;
-    for (var guardian in group.guardians.values) {
+    for (final guardian in group.guardians.keys) {
       if (shardsIterator.moveNext()) {
-        _shards[guardian.peerId] = SecretShardModel(
-          value: shardsIterator.current,
-          ownerId: PeerId.empty(),
-          ownerName: myDeviceName,
-          groupId: group.id,
-          groupName: group.name,
-          groupSize: group.currentSize,
-          groupThreshold: group.threshold,
-        );
+        messages.add(MessageModel(
+          peerId: guardian,
+          code: MessageCode.setShard,
+          payload: SecretShardModel(
+            id: secretId,
+            ownerId: diContainer.myPeerId,
+            groupId: group.id,
+            groupSize: group.size,
+            groupThreshold: group.threshold,
+            shard: shardsIterator.current,
+          ),
+        ));
       }
     }
-  }
-
-  Future<void> distributeShards([Timer? timer]) async {
-    if (_shards.isEmpty) return timer?.cancel();
-    await Future.wait([
-      for (var shard in _shards.entries)
-        sendToGuardian(MessageModel(
-          peerId: shard.key,
-          type: OperationType.setShard,
-          secretShard: shard.value,
-        ))
-    ]);
-  }
-
-  void removeShard(PeerId peerId) async {
-    _shards.remove(peerId);
-    if (_shards.isEmpty) {
-      final updatedGroup = group.completeGroup();
-      await diContainer.boxRecoveryGroup
-          .put(updatedGroup.id.asKey, updatedGroup);
-    }
-    notifyListeners();
   }
 }
