@@ -1,7 +1,6 @@
-import 'package:amplitude_flutter/amplitude.dart';
 import 'package:wakelock/wakelock.dart';
 
-import '/src/core/theme_data.dart';
+import '/src/core/theme/theme.dart';
 import '/src/core/widgets/common.dart';
 import '/src/core/widgets/icon_of.dart';
 import '/src/core/model/core_model.dart';
@@ -42,7 +41,7 @@ class _LoadingPageState extends State<LoadingPage> {
       children: [
         // Header
         const HeaderBar(
-          caption: 'Restore Group',
+          caption: 'Restore a Vault',
           closeButton: HeaderBarCloseButton(),
         ),
         // Body
@@ -50,55 +49,51 @@ class _LoadingPageState extends State<LoadingPage> {
         Padding(
           padding: paddingH20,
           child: Card(
-              child: Column(
-            children: [
-              Padding(
-                padding: paddingTop20,
-                child: Visibility(
-                  visible: controller.isWaiting,
-                  child: const CircularProgressIndicator.adaptive(),
+            child: Column(
+              children: [
+                Padding(
+                  padding: paddingTop20,
+                  child: Visibility(
+                    visible: controller.isWaiting,
+                    child: const CircularProgressIndicator.adaptive(),
+                  ),
                 ),
-              ),
-              Padding(
-                padding: paddingAll20,
-                child: Text(
-                  'Awaiting Guardian’s response',
-                  style: textStyleSourceSansPro416,
+                Padding(
+                  padding: paddingAll20,
+                  child: Text(
+                    'Awaiting ${controller.qrCode!.peerId.name}’s response',
+                    style: textStyleSourceSansPro416,
+                  ),
                 ),
-              ),
-            ],
-          )),
+              ],
+            ),
+          ),
         ),
       ],
     );
   }
 
-  void _showSuccess(MessageModel message, RecoveryGroupModel group) {
-    final controller = context.read<RestoreGroupController>();
-    final count = group.maxSize - group.currentSize;
-    count == 0 // is last Guardian
+  void _showSuccess(MessageModel message) {
+    final count = message.recoveryGroup.maxSize - message.recoveryGroup.size;
+    count == 0
         ? showModalBottomSheet(
             context: context,
             isDismissible: false,
             isScrollControlled: true,
-            builder: (BuildContext context) => BottomSheetWidget(
+            builder: (context) => BottomSheetWidget(
               icon: const IconOf.secrets(isBig: true, bage: BageType.ok),
               titleString: 'Ownership Changed',
               textSpan: [
-                const TextSpan(text: 'The ownership of the group '),
+                const TextSpan(text: 'The ownership of the Vault '),
                 TextSpan(
-                  text: message.secretShard.groupName,
+                  text: message.groupId.name,
                   style: textStyleBold,
                 ),
                 const TextSpan(text: ' has been transferred to your device.'),
               ],
               footer: PrimaryButton(
                 text: 'Done',
-                onPressed: () {
-                  Amplitude.getInstance()
-                      .logEvent('RestoreGroup Finish ');
-                  Navigator.of(context).pop();
-                },
+                onPressed: Navigator.of(context).pop,
               ),
             ),
           ).then(Navigator.of(context).pop)
@@ -107,18 +102,19 @@ class _LoadingPageState extends State<LoadingPage> {
             isDismissible: false,
             isScrollControlled: true,
             useRootNavigator: true,
-            builder: (BuildContext context) => BottomSheetWidget(
+            builder: (context) => BottomSheetWidget(
               icon: const IconOf.secrets(isBig: true, bage: BageType.ok),
               titleString: 'Ownership Transfer Approved',
               textSpan: [
                 TextSpan(
-                  text: message.secretShard.ownerName,
+                  text: message.peerId.name,
                   style: textStyleBold,
                 ),
                 const TextSpan(
-                    text: ' approved the transfer of ownership for the group '),
+                  text: ' approved the transfer of ownership for the Vault ',
+                ),
                 TextSpan(
-                  text: message.secretShard.groupName,
+                  text: message.groupId.name,
                   style: textStyleBold,
                 ),
               ],
@@ -127,19 +123,10 @@ class _LoadingPageState extends State<LoadingPage> {
                 child: Container(
                   decoration: boxDecoration,
                   padding: paddingAll20,
-                  child: RichText(
-                    text: TextSpan(
-                      style: textStyleSourceSansPro416,
-                      children: [
-                        TextSpan(text: 'Get $count more approvals from other '),
-                        TextSpan(
-                          text: message.secretShard.groupName,
-                          style: textStyleBold,
-                        ),
-                        const TextSpan(
-                            text: ' Guardians to finish the restoration.'),
-                      ],
-                    ),
+                  child: Text(
+                    'Get $count more approvals from other Guardians '
+                    'of the Vault to finish the restoration.',
+                    style: textStyleSourceSansPro416,
                   ),
                 ),
               ),
@@ -147,7 +134,10 @@ class _LoadingPageState extends State<LoadingPage> {
                 text: 'Scan New QR Code',
                 onPressed: () {
                   Navigator.of(context).pop();
-                  controller.scanAnotherCode();
+                  Navigator.of(context).pushReplacementNamed(
+                    '/recovery_group/restore',
+                    arguments: true,
+                  );
                 },
               ),
             ),
@@ -158,18 +148,19 @@ class _LoadingPageState extends State<LoadingPage> {
         context: context,
         isDismissible: false,
         isScrollControlled: true,
-        builder: (BuildContext context) => BottomSheetWidget(
+        builder: (context) => BottomSheetWidget(
           icon: const IconOf.secrets(isBig: true, bage: BageType.error),
           titleString: 'Ownership Transfer Rejected',
           textSpan: [
             TextSpan(
-              text: message.secretShard.ownerName,
+              text: message.peerId.name,
               style: textStyleBold,
             ),
             const TextSpan(
-                text: ' rejected the transfer of ownership for the group '),
+              text: ' rejected the transfer of ownership for the Vault ',
+            ),
             TextSpan(
-              text: message.secretShard.groupName,
+              text: message.groupId.name,
               style: textStyleBold,
             ),
           ],
@@ -180,7 +171,7 @@ class _LoadingPageState extends State<LoadingPage> {
               padding: paddingAll20,
               child: Text(
                 'Since Guardian rejected ownership transfer, '
-                'it is impossible to restore the Group.',
+                'it is impossible to restore the Vault.',
                 style: textStyleSourceSansPro416Purple,
                 textAlign: TextAlign.center,
               ),
@@ -197,10 +188,10 @@ class _LoadingPageState extends State<LoadingPage> {
         context: context,
         isDismissible: false,
         isScrollControlled: true,
-        builder: (BuildContext context) => BottomSheetWidget(
+        builder: (context) => BottomSheetWidget(
           icon: const IconOf.secrets(isBig: true, bage: BageType.error),
           titleString: 'Ownership Transfer Failed',
-          textString: 'You are the owner of this Group already!',
+          textString: 'You are the owner of this Vault already!',
           footer: PrimaryButton(
             text: 'Close',
             onPressed: Navigator.of(context).pop,
@@ -208,25 +199,25 @@ class _LoadingPageState extends State<LoadingPage> {
         ),
       ).then(Navigator.of(context).pop);
 
-  void _showError(MessageModel message) {
-    final controller = context.read<RestoreGroupController>();
-    showModalBottomSheet(
-      context: context,
-      isDismissible: false,
-      isScrollControlled: true,
-      builder: (BuildContext context) => BottomSheetWidget(
-        icon: const IconOf.secrets(isBig: true, bage: BageType.error),
-        titleString: 'Ownership Transfer Failed',
-        textString:
-            'We couldn’t finish scanning the QR Code.\nPlease try again.',
-        footer: PrimaryButton(
-          text: 'Scan QR Code Again',
-          onPressed: () {
-            Navigator.of(context).pop();
-            controller.scanAnotherCode();
-          },
+  void _showError(MessageModel message) => showModalBottomSheet(
+        context: context,
+        isDismissible: false,
+        isScrollControlled: true,
+        builder: (context) => BottomSheetWidget(
+          icon: const IconOf.secrets(isBig: true, bage: BageType.error),
+          titleString: 'Ownership Transfer Failed',
+          textString:
+              'We couldn’t finish scanning the QR Code.\nPlease try again.',
+          footer: PrimaryButton(
+            text: 'Scan QR Code Again',
+            onPressed: () {
+              Navigator.of(context).pop();
+              Navigator.of(context).pushReplacementNamed(
+                '/recovery_group/restore',
+                arguments: true,
+              );
+            },
+          ),
         ),
-      ),
-    );
-  }
+      );
 }
